[Refactor] Remove Northstar
